Serial Communications over USB
This is an excerpt from a project which involves taking arguments to the .exe file and sending parametres to an Arduino over USB and saving returned data from the Arduino to a .log file. I have only included the serial communication part of this and the rest was application specific, feel free to copy and manipulate this code to communicate over serial comms with your own USB devices.